Move basic clickable qlabel to utils file
